+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 - Use more type declarations in the tests (#346)
 
 ### Deprecated
+- Initialize `TemplateHelper` lazily (#350)
 
 ### Removed
 - Remove unused `limit` parameter from `DataMapper::findByPageUid` (#348)